Consolidate generateAppSecret reducer with appChange
- generateAppSecret shares appChange 'isFetching' and 'error' fields
- Displaying clientSecret logic moved solely to AppChangeSuccessPage
