+++ app-management/components/AppDetailsPage/AppSecret.jsx
@@ -1,8 +1,6 @@
 import React, { PropTypes } from 'react';
-import LoadingSpinner from '../LoadingSpinner';
-import ErrorAlert from '../ErrorAlert';
 
-const AppSecret = ({ clickHandler, appSecret: { clientSecret, error, isFetching } }) => (
+const AppSecret = ({ clickHandler }) => (
   <div>
     <div className="col-md-8">
       <a href="#client-secret-generation" data-toggle="collapse">Regenerate clientSecret</a>
@@ -12,36 +10,21 @@ const AppSecret = ({ clickHandler, appSecret: { clientSecret, error, isFetching 
         id="client-secret-generation"
       >
         <div className="panel-body">
-          <ErrorAlert error={error} />
-          <LoadingSpinner loading={isFetching} />
-          {clientSecret ? (
-            <p>
-              <strong>
-                Here is your new clientSecret. Make sure to copy it now. You will not be able
-                to see it again.
-              </strong>
-              <br />
-              <code>{clientSecret}</code>
-            </p>
-          ) : (
-            <div>
-              <p>
-                You can regenerate your clientSecret below.
-              </p>
-              <p>
-                <strong>WARNING: </strong>
-                This process will clear your old clientSecret and require updates to your
-                apps or scripts.
-              </p>
-              <button
-                type="button"
-                className="btn orange small"
-                onClick={clickHandler}
-              >
-                Regenerate clientSecret
-              </button>
-            </div>
-          )}
+          <p>
+            You can regenerate your clientSecret below.
+          </p>
+          <p>
+            <strong>WARNING: </strong>
+            This process will clear your old clientSecret and require updates to your
+            apps or scripts.
+          </p>
+          <button
+            type="button"
+            className="btn orange small"
+            onClick={clickHandler}
+          >
+            Regenerate clientSecret
+          </button>
         </div>
       </div>
     </div>
@@ -50,11 +33,6 @@ const AppSecret = ({ clickHandler, appSecret: { clientSecret, error, isFetching 
 
 AppSecret.propTypes = {
   clickHandler: PropTypes.func.isRequired,
-  appSecret: PropTypes.shape({
-    clientSecret: PropTypes.string.isRequired,
-    error: PropTypes.string.isRequired,
-    isFetching: PropTypes.bool.isRequired,
-  }).isRequired,
 };
 
 export default AppSecret;
