+++ app-management/reducers/index.js
@@ -6,12 +6,10 @@ import signupReducer from './signup';
 import appListingReducer from './appListing';
 import appDetailsReducer from './appDetails';
 import appChangeReducer from './appChange';
-import generateAppSecretReducer from './generateAppSecret';
 
 const appReducer = combineReducers({
   appDetails: appDetailsReducer,
   appListing: appListingReducer,
-  appSecret: generateAppSecretReducer,
   auth: authReducer,
   form: formReducer,
   appChange: appChangeReducer,
