+++ app-management/tests/components/AppDetailsPage/AppDetailsPage.test.jsx
@@ -14,17 +14,13 @@ import appFactory from '../../app.mock';
 const store = createStore(appReducer);
 const defaultProps = {
   appChange: {
+    app: {},
+    clientSecret: '',
     error: '',
     isFetching: false,
   },
   appDetails: {
     app: appFactory('id-1'),
-    error: '',
-    isFetching: false,
-  },
-  appSecret: {
-    app: {},
-    clientSecret: '',
     error: '',
     isFetching: false,
   },
@@ -85,11 +81,11 @@ describe('<AppdetailsPage />', () => {
   });
 
   describe('mapStateToProps', () => {
-    it('returns entire appDetails and appSecret states', () => {
+    it('returns entire appDetails and appChange states', () => {
       const state = mapStateToProps(store.getState());
 
       expect(state.appDetails).toEqual(store.getState().appDetails);
-      expect(state.appSecret).toEqual(store.getState().appSecret);
+      expect(state.appChange).toEqual(store.getState().appChange);
     });
   });
